actualizacion de version

+++ final/src/main/java/com/example/demo/Controller/ProductoController.java
@@ -13,6 +13,7 @@ import com.example.demo.Service.ProductoService;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+import java.math.BigDecimal;
 
 @RestController
 @RequestMapping("/api/productos")
@@ -26,10 +27,6 @@ public class ProductoController {
         this.productoService = productoService;
     }
 
-    /**
-     * Obtener todos los productos
-     * GET /api/productos
-     */
     @GetMapping
     public ResponseEntity<List<ProductoDTO>> obtenerTodos() {
         List<EntityProducto> entidades = productoService.listarTodos();
@@ -39,10 +36,6 @@ public class ProductoController {
         return ResponseEntity.ok(dtos);
     }
 
-    /**
-     * Obtener producto por ID
-     * GET /api/productos/{id}
-     */
     @GetMapping("/{id}")
     public ResponseEntity<ProductoDTO> obtenerPorId(@PathVariable Integer id) {
         Optional<EntityProducto> optionalProducto = productoService.buscarPorId(id);
@@ -53,10 +46,6 @@ public class ProductoController {
         }
     }
 
-    /**
-     * Buscar productos por nombre
-     * GET /api/productos/buscar?nombre=
-     */
     @GetMapping("/buscar")
     public ResponseEntity<List<ProductoDTO>> buscarPorNombre(@RequestParam String nombre) {
         List<EntityProducto> entidades = productoService.buscarPorNombreContiene(nombre);
@@ -66,10 +55,6 @@ public class ProductoController {
         return ResponseEntity.ok(dtos);
     }
 
-    /**
-     * Obtener productos por proveedor
-     * GET /api/productos/proveedor/{proveedorId}
-     */
     @GetMapping("/proveedor/{proveedorId}")
     public ResponseEntity<List<ProductoDTO>> obtenerPorProveedor(@PathVariable Integer proveedorId) {
         List<EntityProducto> entidades = productoService.buscarPorProveedorId(proveedorId);
@@ -79,14 +64,10 @@ public class ProductoController {
         return ResponseEntity.ok(dtos);
     }
 
-    /**
-     * Obtener productos por rango de precio
-     * GET /api/productos/precio?min=&max=
-     */
     @GetMapping("/precio")
     public ResponseEntity<List<ProductoDTO>> obtenerPorRangoPrecio(
-            @RequestParam Double min,
-            @RequestParam Double max) {
+            @RequestParam BigDecimal min,
+            @RequestParam BigDecimal max) {
         List<EntityProducto> entidades = productoService.buscarPorRangoPrecio(min, max);
         List<ProductoDTO> dtos = entidades.stream()
                 .map(this::convertirADTO)
@@ -94,7 +75,6 @@ public class ProductoController {
         return ResponseEntity.ok(dtos);
     }
 
-    
     @GetMapping("/mas-vendidos")
     public ResponseEntity<List<ProductoDTO>> obtenerMasVendidos(
             @RequestParam(defaultValue = "10") Integer limite) {
@@ -117,10 +97,6 @@ public class ProductoController {
         }
     }
 
-    /**
-     * Actualizar producto existente
-     * PUT /api/productos/{id}
-     */
     @PutMapping("/{id}")
     public ResponseEntity<ProductoDTO> actualizar(
             @PathVariable Integer id,
@@ -144,10 +120,6 @@ public class ProductoController {
         }
     }
 
-    /**
-     * Eliminar producto
-     * DELETE /api/productos/{id}
-     */
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> eliminar(@PathVariable Integer id) {
         try {
