+++ final/src/main/java/com/example/demo/Repository/VentaDetRepository.java
@@ -3,6 +3,7 @@ package com.example.demo.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.List;
+import java.math.BigDecimal;
 import com.example.demo.Entity.EntityVentaDet;
 
 @Repository
@@ -10,7 +11,7 @@ public interface VentaDetRepository extends JpaRepository<EntityVentaDet, Intege
     List<EntityVentaDet> findByVenta_VentaId(Integer ventaId);
     List<EntityVentaDet> findByProducto_ProductoId(Integer productoId);
     List<EntityVentaDet> findByCantidadGreaterThan(Integer cantidad);
-    List<EntityVentaDet> findByTotalBetween(Double min, Double max);
+    List<EntityVentaDet> findByTotalBetween(BigDecimal min, BigDecimal max);
     List<EntityVentaDet> findByVenta_VentaIdIn(List<Integer> ventaIds);
     List<EntityVentaDet> findByVenta_Cliente_ClienteId(Integer clienteId);
 }
